+++ alphabase/spectral_library/reader.py
@@ -1,14 +1,19 @@
 import typing
+import os
 import numpy as np
 import pandas as pd
 
 from alphabase.peptide.mobility import mobility_to_ccs_for_df
 from alphabase.io.psm_reader.dia_search_reader import SpectronautReader
+from alphabase.io.psm_reader.maxquant_reader import MaxQuantReader
 from alphabase.spectral_library.base import SpecLibBase
 from alphabase.psm_reader.psm_reader import psm_reader_yaml
 from alphabase.psm_reader import psm_reader_provider
 
-class SWATHLibraryReader(SpectronautReader, SpecLibBase):
+from alphabase.constants._const import CONST_FILE_FOLDER, PEAK_INTENSITY_DTYPE
+from alphabase.yaml_utils import load_yaml
+
+class LibraryReaderBase(MaxQuantReader, SpecLibBase):
     def __init__(self,
         charged_frag_types:typing.List[str] = [
             'b_z1','b_z2','y_z1', 'y_z2', 
@@ -20,7 +25,7 @@ class SWATHLibraryReader(SpectronautReader, SpecLibBase):
         fdr = 0.01,
         fixed_C57 = False,
         mod_seq_columns=psm_reader_yaml[
-            'spectronaut'
+            'library_reader_base'
         ]['mod_seq_columns'],
         rt_unit='irt',
         precursor_mz_min:float = 400,
@@ -28,6 +33,49 @@ class SWATHLibraryReader(SpectronautReader, SpecLibBase):
         decoy:str = None,
         **kwargs
     ):
+        """
+
+        Base class for reading spectral libraries from long format csv files.
+
+        Parameters
+        ----------
+
+        charged_frag_types: list of str
+            List of fragment types to be used in the spectral library.
+            The default is ['b_z1','b_z2','y_z1', 'y_z2', 'b_modloss_z1','b_modloss_z2','y_modloss_z1', 'y_modloss_z2']
+
+        column_mapping: dict
+            Dictionary mapping the column names in the csv file to the column names in the spectral library.
+            The default is None, which uses the `library_reader_base` column mapping in `psm_reader.yaml` 
+
+        modification_mapping: dict
+            Dictionary mapping the modification names in the csv file to the modification names in the spectral library.
+
+        fdr: float
+            False discovery rate threshold for filtering the spectral library.
+            default is 0.01
+
+        fixed_C57: bool
+    
+        mod_seq_columns: list of str
+            List of column names in the csv file containing the modified sequence.
+            By default the mapping is taken from `psm_reader.yaml`
+
+        rt_unit: str
+            Unit of the retention time column in the csv file.
+            The default is 'irt'
+
+        precursor_mz_min: float
+            Minimum precursor m/z value for filtering the spectral library.
+
+        precursor_mz_max: float
+            Maximum precursor m/z value for filtering the spectral library.
+
+        decoy: str
+            Decoy type for the spectral library.
+            Can be either `pseudo_reverse` or `diann`
+
+        """
         SpecLibBase.__init__(self,
             charged_frag_types = charged_frag_types,
             precursor_mz_min=precursor_mz_min,
@@ -35,7 +83,7 @@ class SWATHLibraryReader(SpectronautReader, SpecLibBase):
             decoy=decoy
         )
 
-        SpectronautReader.__init__(self, 
+        MaxQuantReader.__init__(self, 
             column_mapping = column_mapping,
             modification_mapping = modification_mapping,
             fdr = fdr,
@@ -45,70 +93,95 @@ class SWATHLibraryReader(SpectronautReader, SpecLibBase):
             rt_unit=rt_unit,
         )
 
-        self._frag_type_columns = "FragmentType FragmentIonType ProductType ProductIonType".split(' ')
-        self._frag_number_columns = "FragmentNumber FragmentSeriesNumber".split(' ')
-        self._frag_charge_columns = "FragmentCharge FragmentIonCharge ProductCharge ProductIonCharge".split(' ')
-        self._frag_loss_type_columns = "FragmentLossType FragmentIonLossType ProductLossType ProductIonLossType".split(' ')
-        self._frag_inten_columns = "RelativeIntensity RelativeFragmentIntensity RelativeFragmentIonIntensity LibraryIntensity".split(' ')
+    def _init_column_mapping(self):
+        """
+        Initialize the column mapping from the `psm_reader.yaml` file.
+        """
+        self.column_mapping = psm_reader_yaml[
+            'library_reader_base'
+        ]['column_mapping']
 
     def _find_key_columns(self, lib_df:pd.DataFrame):
-        def find_col(target_columns, df_columns):
-            for col in target_columns:
-                if col in df_columns:
-                    return col
-            return None
 
-        self.mod_seq_col = find_col(self._mod_seq_columns, lib_df.columns)
+        """
+        Find and create the key columns for the spectral library.
+
+        Parameters
+        ----------
+
+        lib_df: pd.DataFrame
+            Dataframe containing the spectral library.
         
-        self.mapped_peptide_columns = self._find_mapped_columns(lib_df)
+        """
 
-        self.frag_type_col = find_col(self._frag_type_columns, lib_df.columns)
-        self.frag_num_col = find_col(self._frag_number_columns, lib_df.columns)
-        self.frag_charge_col = find_col(self._frag_charge_columns, lib_df.columns)
-        self.frag_loss_type_col = find_col(self._frag_loss_type_columns, lib_df.columns)
-        self.frag_inten_col = find_col(self._frag_inten_columns, lib_df.columns)
+        if 'fragment_loss_type' not in lib_df.columns:
+            lib_df['fragment_loss_type'] = ''
 
-        if self.frag_loss_type_col is None:
-            self.frag_loss_type_col = 'FragmentLossType'
-            lib_df[self.frag_loss_type_col] = ''
+        lib_df['fragment_loss_type'].fillna('', inplace=True)
+        lib_df['fragment_loss_type'].replace('noloss','',inplace=True)
+
+
+        if 'mods' not in lib_df.columns:
+            lib_df['mods'] = ''
+
+        if 'mod_sites' not in lib_df.columns:
+            lib_df['mod_sites'] = ''
 
     def _get_fragment_intensity(self, lib_df:pd.DataFrame):
+        """
+
+        Create the self._fragment_intensity dataframe from a given spectral library.
+        In the process, the input dataframe is converted from long format to a precursor dataframe and returned.
+
+        Parameters
+        ----------
+        lib_df: pd.DataFrame
+            Dataframe containing the spectral library.
+
+        Returns
+        -------
+        precursor_df: pd.DataFrame
+            Dataframe containing the fragment intensity.
+        
+        """
         frag_col_dict = dict(zip(
             self.charged_frag_types, 
             range(len(self.charged_frag_types))
         ))
 
         self._find_key_columns(lib_df)
-        lib_df[self.frag_loss_type_col].fillna('', inplace=True)
-        lib_df[self.frag_loss_type_col].replace('noloss','',inplace=True)
 
-        group_cols = [
-            self.mod_seq_col, 
-            self.mapped_peptide_columns['sequence'], 
-            self.mapped_peptide_columns['charge'],
-        ]
+        # drop all columns which are all NaN as they prohibit grouping
+        lib_df = lib_df.dropna(axis=1, how='all')
 
-        if 'raw_name' in self.mapped_peptide_columns:
-            group_cols.append(self.mapped_peptide_columns['raw_name'])
-
-        col_list_dict = dict([(col, []) for col in self.mapped_peptide_columns.values()])
-        col_list_dict[self.mod_seq_col] = []
+        precursor_df_list = []
 
         frag_intens_list = []
         nAA_list = []
+
+        fragment_columns = [
+            'fragment_mz','fragment_type','fragment_charge','fragment_series','fragment_loss_type','fragment_intensity'
+        ]
+
+        # by default, all non-fragment columns are used to group the library
+        non_fragment_columns = list(set(lib_df.columns) - set(fragment_columns))
+
         
         for keys, df_group in lib_df.groupby(
-            group_cols
+            non_fragment_columns
         ):
+            precursor_columns = dict(zip(non_fragment_columns, keys))
 
-            nAA = len(keys[1])
+            nAA = len(precursor_columns['sequence'])
+
             intens = np.zeros(
-                (nAA-1, len(self.charged_frag_types)),dtype=np.float32
+                (nAA-1, len(self.charged_frag_types)),
+                dtype=PEAK_INTENSITY_DTYPE,
             )
             for frag_type, frag_num, loss_type, frag_charge, inten in df_group[
                 [
-                    self.frag_type_col,self.frag_num_col,self.frag_loss_type_col,
-                    self.frag_charge_col,self.frag_inten_col
+                    'fragment_type','fragment_series','fragment_loss_type',
+                    'fragment_charge','fragment_intensity'
                 ]
             ].values:
                 if frag_type in 'abc':
@@ -139,12 +212,12 @@ class SWATHLibraryReader(SpectronautReader, SpecLibBase):
             if max_inten <= 0: continue
             intens /= max_inten
 
-            nAA_list.append(nAA)
-            for col, col_list in col_list_dict.items():
-                col_list.append(df_group[col].values[0])
+            precursor_df_list.append(precursor_columns) 
             frag_intens_list.append(intens)
-        
-        df = pd.DataFrame(col_list_dict)
+            nAA_list.append(nAA)
+
+        df = pd.DataFrame(precursor_df_list)
+
 
         self._fragment_intensity_df = pd.DataFrame(
             np.concatenate(frag_intens_list),
@@ -160,43 +233,53 @@ class SWATHLibraryReader(SpectronautReader, SpecLibBase):
 
         return df
 
-    def _load_file(self, filename):
-        self.csv_sep = self._get_table_delimiter(filename)
-        df = pd.read_csv(filename, sep=self.csv_sep)
-        self._find_mod_seq_column(df)
-
-        df = self._get_fragment_intensity(df)
-
-        return df
-
-    def _post_process(self, 
-        lib_df
-    ):  
-        self._psm_df['nAA'] = self._psm_df.sequence.str.len()
-        self._psm_df[
-            ['frag_start_idx','frag_stop_idx']
-        ] = lib_df[['frag_start_idx','frag_stop_idx']]
-
-        self.normalize_rt_by_raw_name()
-
-        if (
-            'mobility' in self._psm_df.columns
+    def _load_file(
+            self, 
+            filename:str
         ):
+        """
+        Load the spectral library from a csv file.
+        Reimplementation of `PSMReaderBase._translate_columns`.
+        """
+
+        csv_sep = self._get_table_delimiter(filename)
+
+        df = pd.read_csv(filename, sep=csv_sep)
+        self._find_mod_seq_column(df)
+        
+        return df
+        
+    def _post_process(
+        self, 
+        lib_df:pd.DataFrame,
+    ):  
+        """
+        Process the spectral library and create the `fragment_intensity`, `fragment_mz`dataframe.
+        Reimplementation of `PSMReaderBase._post_process`.
+        """
+        
+        if 'nAA' not in self._psm_df.columns:
+            self._psm_df['nAA'] = self._psm_df.sequence.str.len()
+
+        self._psm_df = self._get_fragment_intensity(self._psm_df)
+        
+        self.normalize_rt_by_raw_name()
+        
+        if 'mobility' in self._psm_df.columns:
             self._psm_df['ccs'] = (
                 mobility_to_ccs_for_df(
                     self._psm_df,
                     'mobility'
                 )
             )
-        
-        self._psm_df = self._psm_df[
-            ~self._psm_df.mods.isna()
-        ].reset_index(drop=True)
 
+        self._psm_df.drop('modified_sequence', axis=1, inplace=True)
         self._precursor_df = self._psm_df
 
         self.calc_fragment_mz_df()
 
+# legacy
+SWATHLibraryReader = LibraryReaderBase
 
 class LibraryReaderFromRawData(SpecLibBase):
     def __init__(self,